+++ ddlx.sql
@@ -393,6 +393,13 @@ AS $function$
          null as acl
     FROM pg_subscription sub
    WHERE sub.oid = $1
+   UNION ALL
+  SELECT e.oid, 'pg_extension'::regclass,
+    e.extname AS name, e.extnamespace::text AS namespace, pg_get_userbyid(e.extowner) AS owner,
+    'EXTENSION'::text AS sql_kind,
+    e.extname AS sql_identifier,
+    NULL::aclitem[] AS acl
+   FROM pg_extension e;   
 #end
 $function$  strict;