+++ ddlx.sql
@@ -1,6 +1,6 @@
 --
 --  DDL eXtractor functions
---  version 0.21 lacanoid@ljudmila.org
+--  version 0.22 lacanoid@ljudmila.org
 --
 ---------------------------------------------------
 
@@ -2905,7 +2905,8 @@ select row_number() over() as n,
   from ddlx_get_dependants($1) gd
   left join pg_depend de
        on de.objid=gd.objid and de.refclassid='pg_extension'::regclass
- where 'ext' ilike any($2) or de.refclassid is null
+ where ('ext' ilike any($2) or de.refclassid is null)
+       and gd.classid not in ('pg_amproc'::regclass,'pg_amop'::regclass)
 )
 select ddlx_create($1,$2) as ddl_create,
        ddlx_drop($1,$2) as ddl_drop,