+++ CHANGELOG.md
@@ -3,8 +3,8 @@ Version 0.21
 - improvements to `ddlx_create_collation`
 - added `IF NOT EXISTS` in a few more places
 - new options `nodcl`,`noowner`,`nogrants`,`noalter`
-- `ddlx_identify` now identifies extensions
-- addded `ddlx_create_extension()` function
+- `ddlx_identify` now identifies extension and added `ddlx_create_extension()` function, so
+- extensions are now supported
 - owner now dumped only if distinct from current role or option `owner` is specified
 
 Version 0.20